Update Main.c fixes major problem
As it turns out I forgot to reinitialize the screen after power off, so when I finally got the PCBs and soldered the PIC onto it and was testing it, I kept running into problems with the device resetting.  Turns out I was not initializing the screen after restarting it.